Update README.md
Please add some projects of FLUTTER to this list.
Flutter is a UI toolkit for building fast, beautiful, natively compiled applications for mobile, web, and desktop with one programing language and single codebase. It is free and open-source. Initially, it was developed from Google and now manages by an ECMA standard. Flutter apps use Dart programming language for creating an app. Flutter can be used to develop cross platform applications for Android , iOS and web. The latest update on flutter reveals that it can be used to develop desktop applications for windows.
The official documentation for flutter is available at  - https://flutter.dev/