Add Scribe-iOS to data.json
Hi there! 

Hoping that we can be added to this list. Scribe is still relatively new, but we're slowly making headway towards v2.0 that adds autocomplete and autosuggest - the last two must have features for a wider audience. The app's data is mostly from [Wikidata](https://www.wikidata.org/wiki/Wikidata:Main_Page) and Wikipedia, so we're a [featured project](https://www.mediawiki.org/wiki/New_Developers) for developers who are new to Wikimedia :)

Thanks for this list!