Cleaned up XWiki entry
Going against the rules (already) this links to the issues tracker as otherwise it's not immediately clear how to find these. The link is no longer broken though as the link to a query for the label didn't seem to be working.
Re-added the link for the onboarding guide from the original pull request.